feat: Enhance backend utils structure
