+++ apps/backend/src/utils/errors/index.ts
@@ -1,0 +1,1 @@
+export * from "./errors";
