+++ apps/backend/src/utils/helpers/filterHelper.ts
