+++ apps/backend/src/utils/helpers/generateRoutes.ts
@@ -1,5 +1,5 @@
 import { Express } from "express";
-import * as AdminRoutes from "../routes/admin";
+import * as AdminRoutes from "../../routes/admin";
 
 const adminRouteList = [{ path: "/auth", router: AdminRoutes.adminAuthRouter }];
