+++ apps/backend/src/utils/helpers/utils.ts
